rust: Fix multiple rlib conflict when building with rust and libstd-rs
When building an image with IMAGE_INSTALL:append = ' rust libstd-rs',
a conflict arises due to multiple candidates for the rlib
dependency `std`. This results in the following error:

error[E0464]: multiple candidates for `rlib` dependency `std` found
  |
  = note: candidate #1: /usr/lib/rustlib/x86_64-poky-linux-gnu/lib/libstd-20c3de2d9292cd03.rlib
  = note: candidate #2: /usr/lib/rustlib/x86_64-poky-linux-gnu/lib/libstd.so

The issue seems to be from an extra copy of the rlib as both the recipes
generate the same rlibs, causing conflicts when using std lib to compile
rust programs in the image.

Remove the redundant rlib copy ensuring only the necessary rlib is
present and prevent the conflict.

Signed-off-by: Yash Shinde <Yash.Shinde@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>